Chore: refactor
